requirements.txt
python-dotenv==1.0.1
playwright==1.44.0
httpx==0.27.0
anthropic==0.28.0
google-generativeai==0.7.2
sqlalchemy==2.0.30
jinja2==3.1.4
resend==2.1.0
pytest==8.2.2
pytest-asyncio==0.23.7
python-dateutil==2.9.0
fastapi==0.115.6
uvicorn[standard]==0.32.1

